back-end/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
anthropic==0.40.0
langchain==0.3.3
langchain-anthropic==0.3.0
langchain-core==0.3.21
pydantic==2.10.3
